+++ aiu/youtube.py
@@ -5,7 +5,9 @@ from typing import TYPE_CHECKING
 from urllib.parse import parse_qs, urlparse
 from tqdm import tqdm
 from youtube_dl import YoutubeDL
-from ytm import YouTubeMusic, YouTubeMusicDL  # noqa
+from ytm.apis.YouTubeMusic import YouTubeMusic
+from ytm.apis.YouTubeMusicDL import YouTubeMusicDL
+from ytm.types.ids.ArtistId import ArtistId
 from ytm import utils as ytm_utils
 
 from aiu import LOGGER
@@ -13,7 +15,7 @@ from aiu.parser import fetch_image
 from aiu.typedefs import Duration
 
 if TYPE_CHECKING:
-    from typing import Optional, Tuple
+    from typing import Dict, List, Optional, Tuple, Union
     from aiu.typedefs import JSON
 
 
@@ -121,7 +123,7 @@ class TqdmYouTubeMusicDL(CachedYoutubeMusicDL):
 
 
 def get_reference_id(link):
-    # type: (str) -> Tuple[bool, bool, str]
+    # type: (str) -> Tuple[bool, bool, Union[str, List[str]]]
     """
     Finds the appropriate reference ID from a YouTube Music/Video link.
 
@@ -137,10 +139,13 @@ def get_reference_id(link):
         raise ValueError("Invalid YouTube Music/Video link located at invalid host: [{!s}]".format(link))
     query = urlparse(link).query
     params = parse_qs(query)
+    # format: <youtube-link>/watch?v=<ID>
     if music_link and not any(ref in params for ref in ["v", "list"]):
         raise ValueError("Invalid YouTube Music link does not provide a song or album reference: [{!s}]".format(link))
     elif video_link and "v" not in params:  # ignore list (video playlist)
         raise ValueError("Invalid YouTube Video link does not provide a video reference: [{!s}]".format(link))
+    # format: <youtube-link>/playlist?list=<ID>
+    # note: "list=<ID>" can also be in "watch?v=<ID>" format
     if music_link and "list" in params:  # process list first in case somehow both watch/list are present
         album = params["list"][0]
         LOGGER.debug("Found YouTube Music album ID: [%s]", album)
@@ -152,6 +157,44 @@ def get_reference_id(link):
     video = params["v"][0]
     LOGGER.debug("Found YouTube Video song ID: [%s]", video)
     return False, False, video
+
+
+def get_artist_albums(link, throw=True):
+    # type: (str, bool) -> List[Dict[str, str]]
+    """
+    Obtains all album IDs produced by a given artist ID extracted from appropriate YouTube Music link.
+    """
+    try:
+        parts = link.split("/channel/")
+        if len(parts) != 2:
+            raise ValueError(f"Not a valid channel link: [{link}]")
+        artist = ArtistId(parts[1])  # raise TypeError if invalid
+    except (TypeError, ValueError):
+        if throw:
+            raise
+        return []
+    api = YouTubeMusic()
+    meta = api.artist(artist)
+    albums = meta.get("albums", {}).get("items", [])
+    # get the playlist ID instead of Album ID to form the corresponding download/listing YouTube Music links
+    album_meta = [
+        {
+            "name": info["name"],
+            "link": parts[0] + "/playlist?list={}".format(info["shuffle"]["playlist_id"]),
+            "id": info["shuffle"]["playlist_id"],
+        }
+        for info in albums
+    ]
+    # duplicate album names is allowed (usually duplicate uploads, contents equivalent)
+    # remove them since only unique output directories can be created
+    album_found = []
+    album_names = set()
+    for album_info in album_meta:
+        if album_info["name"] in album_names:
+            continue
+        album_names.add(album_info["name"])
+        album_found.append(album_info)
+    return album_found
 
 
 def update_metadata(meta, fetch_cover=False):
@@ -237,9 +280,9 @@ def fetch_files(link, output_dir, with_cover=True, show_progress=True):
             covers = meta["thumbnails"]
             for img in covers:
                 img["ratio"] = abs((float(img["width"]) / float(img["height"])) - 1)
-            covers = list(sorted(covers, key=lambda img: img["ratio"]))
-            covers = list(filter(lambda img: img["ratio"] == covers[0]["ratio"], covers))
-            covers = list(sorted(covers, key=lambda img: img["height"]))
+            covers = list(sorted(covers, key=lambda _img: _img["ratio"]))
+            covers = list(filter(lambda _img: _img["ratio"] == covers[0]["ratio"], covers))
+            covers = list(sorted(covers, key=lambda _img: _img["height"]))
             thumbnail = covers[0]
         if isinstance(thumbnail, dict):
             url = thumbnail["url"]